Raise Daemon priority while exiting
A zygote process has to wait until Daemons have actually exited
at the OS level before forking. There is some suspicion that
there can be a significant delay between the thread becoming
joinable and the thread actually disappearing. Try to address
that by increasing the thread priority just as it's about
to exit.

The now higher priority thread should not run for long enough
to matter. But if it doesn't run, it can block the zygote's
critical path.

Bug: 314247665
Bug: 314225311
Bug: 304183774
Bug: 310106506
Test: Build and boot AOSP
Change-Id: I1d107b7a743a233660aec5606a1ccb0c069e2719