Reland "File.getCanonicalPath() should remove parent directory /.. at the rootfs"
File.getCanonicalPath() can return /../<file> when the realpath(3)
doesn't resolve the path for a likely non-existing file. It
can be further canonicalized to be /<file> as /.. is / at the rootfs.
As the API doc specifies that the method should return a unique
canonical form, this CL is the best effort to return a unique form.

Bug: 312399441
Test: CtsLibcoreTestCases CtsLibcoreOjTestCases
Change-Id: I6e39b0b1cac8e01cd5e73346592a16ca4e66a745