Add native API call VMRuntime.getFullGcCount()
Bug: 331243037

To report memory metrics for native allocations, a point in time
is needed where a full GC has been done and all objects on
ReferenceQueue have been cleaned up.

This CL adds the API of `getFullGcCount()` to `class VMRuntime`
so by comparing the full GC count with the previous one, it knows
if a full GC has happened.

Change-Id: I9e178389c76b72636c6aaae1f4a861f2904225db