Use VarHandle for counting native allocations
Bug: 331243037
Flag: EXEMPT refactor

The current NativeAllocationRegistry API requires objects with different
sizes to have its own NativeAllocationRegistry. This means allocations for
each registry will have the same size.

This change uses VarHandle to only keep the counter of the allocations
for each registry, and the total size in bytes is then calculated when
getMetrics() is called.

This reduces the size of each registry, and improves performance by avoiding
sync-lock every time updating metrics.

Change-Id: I0c8cb4b2d8195c630d1121cd469da5024287c9de