Remove the dependency on java.nio.Bits from HeapByteBuffer
The indirect read and write via Bits is unncessary. Inline the byte
read / write in HeapByteBuffer.

The micro-benchmark is similiar before and after this CL on x86_64.

If I actually inline putFloatUnchecked(int, float) in putFloat(float),
and update the boot-image-profile.txt, the micro-benchmark is slight
improved:

android.libcore.regression.ByteBufferPerfTest_timeByteBuffer_putFloat[5]
Before -
        timeByteBuffer_putFloat_mean (ns): 4645
After -
        timeByteBuffer_putFloat_mean (ns): 4351

Note that a temporary regression is introduced, because the methods, e.g.
HeapByteBuffer.putFloatUnchecked, putIntUnchecked, are not compiled
until boot-image-profile.txt is updated by the ART cloud service.

Bug: 362159978
Test: m droid
Change-Id: Ibc61384413b81052d835c15a8fa70f143b3a4f33