Disable InvalidPatternSyntax
Some of the invalid regexes in the tests cause java to throw
an exception that contains a null byte. The null byte is then
printed by errorprone to the build log, causing some tools to
think the build log is a binary file. Disable instead of demoting
to a warning.

Bug: 253827323
Test: Presubmits
Change-Id: Iee94afdb9082f6c269dda1e036892516f943326e